fix: turn off CLI config parsing for service start
On every command line command we run a global function that inits state
for that command line. Most of the time this is okay because every
command line item needs to parse configuration in the same way.

But the `service start` command is special since it needs to parse
things that are needed by the API. So previously when we used --config
on the `service start` command what we would get is instead the program
attempting to parse the service configuration file as the CLI config
file, leading to incorrect parsing failures.

Instead we've made the init function skip the CLI parse if the
`service start` command is the one being run.

Fixes #4