fix: add encryption_key validation
Previously, the encryption key would only fail at runtime, leading to
the possibility that users would encounter an error that required a
restart.

Added a check for the correct length of the encryption_key and also
a check to make sure production runs of the encryption_key are not
set to the default.