feat: Add Github trigger
The github trigger is the proof of concept trigger for Gofer, it's
integration into the system shows whether the system actually makes
any sense and is easy to reason about. Through writing it, it has
improved the story for a lot of how other things within Gofer is run
including the new secrets store and the eventual export feature.

Along with this feature is a new way of passing events through the
event handler and proof that the event handler actually works.