feat: add secretStore interface and default impl
Previously, there was no way to include secrets in your
pipeline configuration files. This was a purposeful decision
to eliminate the need to have the user connect a secret store backend
and keep the program simple.

Through implementing the github plugin I've realized that the idea
of not including a secret store is actually more difficult than just
including one. I think this makes things easier for the user to grok
and makes developing trigger features a bit easier. It is also less
awkward to enter docker registry secrets because users will already
have the concept of a pipeline secret.

This change implements the new interface, the default bolt
implementation, the new documentation changes, the cli tooling,
and removes the docker registry feature.