feat: switch github trigger to github app
This makes the use of the Github trigger a little bit easier and
straightforward by removing the need for the user to supply a
webhook secret. Instead, the hard part is done on the ops side
and all secrets are set up on service start.

This allows the user to simply type in a repository and an event
and get going. And allows the ops teams to have the ability
to have proper centralized permissioning control.

Also update documentation for the Github trigger and the external
events endpoint.