fix: invalid pipeline config resolution
Previously if you had a misconfigured trigger Gofer would not complete
the trigger subscription, but still leave the pipeline as if it had.

This changes the functionality to instead disable the pipeline on create
if there is an error and for create/update remove the trigger from
your pipeline configuration.