fix: add pipeline trigger subscription states
Previously when Gofer had an issue with communicating with a trigger
on behalf of a pipeline configuration it would remove the trigger
from the pipeline.

This was bad due to user confusion and just generally not being explicit
with what is happening.

This adds pipeline trigger states which can be used to convey to the
user that one of their pipeline triggers might be disabled because
of an error.

This also cleans up the spots where we previously would hard fail
for trigger subscription errors. Now we simply mark the subscription
as disabled, disable the pipeline so the user is alerted, and continue
on with the process.

This allows pipelines to be in inbetween states, but also makes it easier
for users to understand what needs fixing.