BREAKING CHANGE: eventbus internal refactor
Events are now managed from a centralized event bus that different
parts of the program can listen to and publish to.

This is a massive internal refactor changing elements in the storage,
api, internal layers.

The event bus also allows outside programs to operate upon Gofer events.

This event bus architecture gives us a platform for implementing
notifications.