BREAKING CHANGE: Move triggers to gofer-containers
The monorepoing of the trigger container for Gofer caused IDE issues
and just generally felt like they didn't belong in the main repository
(which makes sense because they are snippets of code that are tangentially
related, but not quite owned by the main repository).

To alleviate this, I've put them in a separate repo and maybe eventually
we'll create a documentation site so that people can view containers
from this repository as it helps with running things within Gofer.