feature: [part 1] allow users to run shell cmds
The use case being solve here is the very popular model of being
able to define what docker container you want and immediately
being able to run some commands within that docker container.

For example, it is probably a better user experience to instead of
creating your own docker container for each step to instead be able
to initiate small workflows from already pre-created containers.

There are other advantages here too as it lowers the bar to entry
and also allows you to troubleshoot some of your custom containers.

To do this we pass the commands within the script attribute in as a
program and run that program with the given shell the user has defined.