feature: Notifiers
This commit introduces the concept of notifiers to Gofer.

The use case is that most times after the run of a pipeline you will
want to inform some other service or use about the result of said run.

Notifiers allow this functionality by allowing you to run special
pre-registered containers as the last thing in your pipeline.

Notifiers are implemented as pre-registered objects only because
it would be annoying to have each user who wanted to set up a
Slack notifier have to go retrieve the credentials for that particular
notifier before being able to use it.

The implementation of notifiers differ from triggers only at runtime
While triggers are launched and then communicated with, notifiers
are simply run after the pipeline is run. This gives notifiers the
ability to fail the run and also give the user the impression that
the notifier is part of their pipeline.

The downside of this implementation is that notifiers cannot keep
state and therefore need to leverage other options for doing so.

This commit also introduces the concepts of:

* Run lifetime Gofer API tokens
    * This gives the user the ability to use the Gofer CLI or API
    in order to communicate with Gofer without having to supply a
    key for every job.
* New containers layout
    * The previous container URL used underscores instead of simple
    slashes.
* Fixed bug in Publish where events could be potentially published
out of order
* Various CLI improvements.