Add ability to install trigger/notifiers
Trigger and notifier installations previously took the user out of
the flow of using Gofer and was against the Gofer ethos of being
stupidly easy to use.

The typical installation pattern previously for the Github trigger for
example was:

1) Go to documentation
2) Not understand what you were doing and try to follow along
3) Create a bunch of github specific things you aren't familiar with
4) Base64 some keys
5) Take everything that you collected over several long steps and then
throw them into the configuration file and hope the next startup worked.

This is a sticking point for anyone who wants to get this up and running
fast.

With installers the trigger can now guide you through the installation process
itself, making the calls on your behalf and then finally making a call
back to Gofer to install that specific trigger.

Hopefully the new pattern is:

1) Type in `gofer trigger install cron <image>`
2) Follow on-screen instructions (we can now automagically do things like
auto-prompt for oauth and auth to what we need as the user)
3) Done!

* This commit also removes the monitoring triggers goroutine
* Rethinks a bit of how docker interactions should work
* and much more that I don't remember