feat: Refine the ability to add tokens to tasks.
Gofer previously had the ability to auto-generate and inject
API tokens to your tasks. It would do this for every run no matter what.

This is not ideal since the user might be running a container that is
less than 100% trusted and the API Token gives full read access to
the Gofer API.

Instead we've made it so that the pipeline config can specifically
mention which tasks it would like injected with an auto-generated token.