bugfix: make namespace determination more robust
Previously for many API endpoints we would blindly accept
whatever the user's preferred namespace was. This caused downstream
errors sometimes when the namespace was not found in the database.

Created a new verification function to make sure that the intended
namespace exists first and passes an error back to the user.