refactor: Refactor the config package
Previously the configuration for the app was home grown
it used many different config libraries to create a hierarchy
of configuration.

This worked well, but there was a lot of hacks to make up for little
places where the many libraries did not fit perfectly(namely,
the hcl parser code was kinda shit and we had struct tags for
every library)

To fix this and make the configuration code much simplier I swapped
the config package to use https://github.com/knadh/koanf which
greatly simplifies a lot of hoemgrown code. HCL time.Duration parsing
works right out of the box where previously there was a lot of code
to make it work.

The tradeoff is that envs now need to be separated by double underscores
in order to be understood properly. Which can be a confusing interaction
but the maintainabiliy boost is probably worth writing a bit of
extra documentation.