feat: Include up to date env vars in help
Previously we manually wrote down the env vars from the configuration
package. This was a surefire way to have them be inaccurate as they changed
in the code, but was not updated in the documentation.

Because of the change to use th koanf config package that enabled us to
leverage reflect to instead auto-generate what the environment variables
actually are.

This commit makes it so both help commands have auto-generated env vars.