refactor: Removing common tasks
Previously common tasks were meant as a way to allow Gofer administrators to pre-setup
containers and their variables so that it was easy for downstream users to add on to
their pipelines very easily. For example, instead of having to re-input the variables
or credentials for something like the ability to post to Slack, instead the administrator
would pre-set it up with the required variables and then users could simply add that container
to their pipeline.

After thinking about it for a while this was not in the right design direction for Gofer for
several reasons:

* Gofer's pipeline configuration design is meant to allow users to build on top of it. This means
that if you have users who all want a similar container you can simply break that set of configurations
into a separate function or library that your users can use. This allows the concept of "common tasks"
without having to treat them as a special feature.
* The update to global secrets allows for custom tasks to be set-up in such a way that admins can
define beforehand.
* Supporting two task types involved a lot of polyformistic solutions to what should be simple problems.
* The complexity in code eventually lead to complexity that was passed on to the user/admin. The more features
the harder it is to keep track of everything.