Updating how extension installations work
Previously extension installations were done through a process where
the extension would pre-write installation instructions and questions
and then Gofer would pass these along to the user to interpret.

This wasn't great because:
  - It was confusing for extension authors to write these instructions
  - There was no interactivity when installing. Users would follow the
  instructions but get no feedback on if the configuration they were
  putting in was in the correct direction.

Instead the new process is to use Gofer as a middle man to connect
the user and the extension such that the extension can now intelligently
interact with the user and both installing extensions and subscribing
to them is easier.