fix: Bootstrap token logic fix
Previously we had a bug where the bootstrap token couldn't be created on first Gofer boot up.

The previous logic for the bootstrap token route simply checked if any other tokens existed, which broke when the feature of automatic extension tokens was introduced.

The real solution here should be inserting into the DB once we create a bootstrap token, but for now we simply filter out any extension tokens and check to see if there are any remaining tokens.