fix: Extensions now correctly pass registration configurations
Previously because of the rewrite there was a bug where it was impossible to add extensions.

This was due to the fact that we changed the way extensions worked. Instead of having an init function
extensions now work as how they did in their inception which is we start an extension and then wait
to be able to hit a healthz type endpoint.

This lack of an init function means that we lost the code that was previously geared toward passing extension
configuration to extensions.