Improve gofer frontend
The Gofer frontend is often the first touch point a user sees just to make sure the service is upo and operating.

This can make the service feel like it's alive and being used, so I've implemented
a basic frontend just so some thing is there.

It is very very simple and does the bare minimum to show the user that Gofer is alive and working.