refactor: Tokens now support usernames
In the last commit Initiator got tied to tokens by id so that we could determine which user performed what action on runs.

To further support that, instead of just returning the token_id which humans don't care about we added a user field for tokens.
This gives the ability for users to have a form of identification that is human readable so that can be reflected in the frontend.

This also gives the advantage for extensions, special tokens, and more to all self-identify so that it's easier and more straightforward
to tell which thing is performing which action.