refactor: Make Shepard operate in more of an event driven fashion.
Previously, we used a mixture of event driven design and shared memory to determine which task executions should
run and when. This was due to trying to keep things simple, but once we started adding more functionality to
this very complicated set of actions it was clear that commiting to the event driven model was somewhat easier
than trying to share the limelight between the two.

This change removes some of the old memory sharing and instead fully commits to the event driven style. We use Gofer's
included event bus to pass messages that control when tasks have started and what status they're currently in.

This allows us to orchestrate all tasks based on events that are currently happening within the Gofer system.

This change will also make it easier to do two things:

1. It will make it so that recovering tasks is easier. Since we can simply launch another `monitor` and that monitor
should be able to query for what happened to the task to appropriately backfill statuses.
2. We should be able to do things like cancellations much easier now. Since every task execution is listening to the
event stream, we no longer have to make sure we're polling to see if the task should be cancelled, instead we just
listen to events and then send a cancellation when we read that we should have one. This also makes it so that
we don't have the leaky abstraction we had earlier at the scheduler level, such that the scheduler has to differ between
when a container is just normally stopped and a container is "cancelled".