feat: add RBAC permissioning
Previously Gofer tried to subsist on a simple permissioning system since there were more important things.

As progress was being made I realized that we needed finer grained permissioning to better support extensions
which should have access to the Gofer API, but not too much access.

That set of a chain reaction of changes in which I thought that we should just pull the trigger on an RBAC system and
the rest is history. It's not optimized just yet, but it works.

* Old permissioning system (management vs client) has been removed in favor of full RBAC.
* Docs updated.
* Roles subcommand added to the CLI.
* Tokens now support many roles.
* Permissions routes added to the API.