refactor: [Step 1] Allow extensions to use object store.
This is the first step in a likely three part change where we will allow
extensions to use the Gofer object store as their database.

This change comes about due to reliabiliy concerns about Gofer having to be the
one manging state for a sometimes transient extensions. This model
will allow extensions(which are just containers) to crash and be
restarted by the container orchestrator without Gofer having to be
involved in it's restart process.

Currently, Gofer has to be the middle man for all extension startup
actions. This makes it so containers can simply save some of their
state and even give the added boon of Gofer not having to be always
repopulate all pipeline subscriptions on every run. Which also has
the added benefit of making so that extensions can design themselves
around avoiding the thundering herd problem.