feat: [3/4] Create extension subscription endpoint
In order to facilitate extensions repopulating their state on startup
it would be helpful if an extension upon init could simply just ask
who was subscribed to it. To that end added a new subscription list
endpoints specifically for extensions.

This should in theory give most trivial extensions the data they need
to be fully self-sufficent upon startup.

This commit also adds the code to remove the extension role if it
is ever unregistered, preventing errors from reinstalling and generally
just keeping the database clean.