feat: [1/2] Add extension debug endpoint
Extensions have a debug endpoint that helps when troubleshooting.

Exposing that endpoint to admin users with this PR.

This has to be broken between two different PR because there is a
bug in proginator(automatic openapi code generator) where it is
confused on which struct to use to attempt to deserialize the DebugResponse payload.