fix: Run recovery is back
Run recovery was previously one of the broken things left over from the
transition over to rust. During this transition we also redesigned the
event system and "run shepard" so it could better handle different
scenarios. Part of this was making it more event driven.

To that end if Gofer crashes while executing a run it can now correctly
determine where it left off at and pick right back up.

This also includes a forgotten version bump to 0.8.1