Update Github extension to support checks API
Previously we had put the Github extension on the back burner
to focus on other things during the rewrite.

Now that the rust rewrite is over the work to make sure external
extensions worked and that the Github extension was a good
steward of those types of extension is underway.

This commit fixes some obvious bugs in the external event API
and the github extension. It also adds a new event type to the
Github API and the ability for Gofer to report it's runs back
to Github.

This was table stakes for any workflow engine and now that it has
landed it enables Gofer itself to transition into using Gofer.