feat: Allow Gofer to accept anonymous users for certain endpoints.
One of the attributes that attracts people to apps is the "lived-in"
effect. The notion that if users see that the software is being
actively used they are more likely to try it themselves.

This change introduces a anonymous role that allows gofer to serve
read only requests for a subset of endpoints only under the default
namespace. Since the default namespace is the starting point for most
folks this should instantly give visual feedback that gofer is working
as intended.

This also makes it easy for people who are simply poking around with
Gofer to see that there is indeed responsiveness.