Add dogfood dir
This is the first pass at making Gofer use Gofer.

Because Gofer does not control the backend to how services get run it cannot
precache things like github repos. To mitigate this, I beleive it can be
a first class or second class feature for Gofer to automatically build
and update your repo within a container and then just keep that container
updated.

Still deciding on how the UX around this should be designed.k