Fix: Inconsistent checking of ids between underscore and hypens
As Gofer's frontend was further flushed out we realized that user
created IDs would be better represented as alphanumerical strings with
hypens as the only special character allowed. This is because previously
the only special chracter allowed was an underscore, which _should_ work
but goes against commonly accepted web conventions.

This PR should fix numerous bugs that weren't cleaned up in the transfer
between the two.