fix: Thread blocking deadlock within run_shepard
We only found this due to running on a single core system.

Previous to this PR I was using `std::sync::Barrier` which is a mechanism
to make threads wait until all threads have checked in with the `barrier`.

It's a helpful coordination mechnism to make sure the we don't run into
weird race conditions when using many threads. Unfortunately, I used
the std version instead of the tokio version which blocks the thread
in order to wait for the other threads.

This meant that only on systems where the thread count was low enough
(tokio spawns a thread per CPU (I believe)) we were deadlocked whenever
a new run started executing, as the monitor task would come up first
block the executor and then nothing else would be able to run.

tokio-console was very helpful in being able to solve this without
spending more time adding tracing.