feat: CLI now shows exact time for detail conf setting
Previously the CLI would only show an approximate, friendly time for a
certain context. This has been enhanced now to work with the command line
setting `detail`. Enabling this setting will now cause Gofer to print
exact timestamps in all areas where previous approximations were given.

Additionally, this includes a minor refactor for the function that is
used to glean this time.