Add new sdk field to tasks that allow users to always pull images
Previoulsy if you had a container image that had a tag that was consistent
but the image was constantly updated it was impossible to tell Gofer to
check for a new version of the image(This behavior mostly occurs with image
tags like "latest").

This PR add functionality for telling Gofer to pass along the request
to always check for the latest version of the image to the downstream
scheduler.