fix: Update gofer frontend to use a table instead of div blocks.
The div blocks for the Gofer frontend were great to look at but not very
functional. Since all the frontend does is show runs it's better to just
have a table of runs. This also sets the stage for a search function
and sorting functions which would be useful for users simply looking
to explore what has been run already.