Refactor: Improve the ergonomics of the SDKs
Adds some small feature changes that should allow users of the SDK to
be more expressive and less limited by coding limitations.

The largest example of this is the introduction of a "script"
method which allows the user to write in bash commands like they would
normally in a script and gets executed as a script.

The variables method has also been largely changed, getting rid of the
single variable and changing variables to take in a more ergonomic
list of key-value pairs.

This relieves the user, going from something like this:

`.variables(HashMap::from(["Hello".to_string, "World".to_string]))`

to being able to simply write:

`.variables(vec!["Hello", "World"])`