refactor: object store now accepts streaming upload/download
Previously the object store was very simple and as such simply read
entire objects into Gofer's memory before transferring them to the
sqlite database we used as a temporary default object store.

While dogfooding this turned out to be a blocker which restricted files
to a certain size or else Gofer would OOM.

The commit refactors a lot of the Gofer endpoints around objects and
the internal representation of how objects are handled.

Additionally, the sqlite default object store was swapped out for a new
default that uses the filesystem directly instead.

This should enable very large files(currently the limit is 50GB) to be
streamed back and forth when a user or container might need without
endangering the stability of the service.