fix: Resolve bug in stringify logic for cli object cmds
The stringify logic(which may not be needed?) that enables users
to use the object store as a simple key/value string store had a
bug in the logic which checked out large the bytes were before
inserting it into the buffer.

This changes resolves that error and also adds some more documentation
around these commands.