fix: Resolve incorrect logic for object overwrites
Previously we were checking for conflicts before inserting into our
main database when putting into an object store. This is incorrect
because a user can pass force to overwrite an object.

The logic here should be that we still attempt to enter it into the
database(because the user should be able to pass force on an object
not in the store yet), but when that fails we don't complain.