fix: More interpolation woes with variable interpolation
Probably should have written tests for this from the get go and now
I'm paying for it.

Weeding out more bugs in the env var interpolation layer where too many
or not enough brackets ends in despair.

This should fix a long standing bug that existed in the parse_interpolation_syntax
function.