refactor: Turn the run shepherd into the orchestrator
Previously the code that managed how runs are handled was unbounded
and simply a function that got called to trigger a run.

This change turns that into a service of sorts which the main thread
can call to launch a run. This new `orchestrator` has a concurrency
limit backed by semaphores so it can better report on what is going
on at any given time to the main API.